add file CRUD

+++ FileStorage.BLL/DTO/FileInfoDTO.cs
@@ -11,10 +11,11 @@ namespace FileStorage.BLL.DTO
     {
         public string Id { get; set; }
         public string FileName { get; set; }
-        public int Size { get; set; }
+        public double Size { get; set; }
         public bool IsPrivate { get; set; }
         public DateTime DownloadDate { get; set; }
         public string RelativePath { get; set; }
+
         [DisplayName("User")]
         public string UserId { get; set; }
     }
